:recycle: [#15] move externeverwijzing model validators
