+++ src/open_producten/producttypen/models/zaaktype.py
@@ -1,11 +1,10 @@
-from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
 from open_producten.utils.models import BaseModel
 
-from .externeverwijzingconfig import ExterneVerwijzingConfig
 from .producttype import ProductType
+from .validators import check_zaaktypen_url
 
 
 class ZaakType(BaseModel):
@@ -35,10 +34,3 @@ class ZaakType(BaseModel):
 
     def __str__(self):
         return str(self.uuid)
-
-
-def check_zaaktypen_url():
-    if not ExterneVerwijzingConfig.get_solo().processen_url:
-        raise ValidationError(
-            _("De zaaktypen url is niet geconfigureerd in de externe verwijzing config")
-        )
